refactor(common): SessionConst -> SessionConstant로 변경, 패키지는 constant 패키지로 옮김
클래스 이름을 일관되게 줄여쓰지 않고 표현하고 패키지는 좀 더 쉽게 찾을 수 있게 하기 위하여 변경
